+++ Module.php
@@ -8,13 +8,19 @@ use yii\helpers\Inflector;
 
 /**
  * Module
- *
+ * 
+ * @property string $mainLayout Main layout using for module.
+ * Default to layout of parent module.
+ * Its used when `layout` set to 'left-menu', 'right-menu' or 'top-menu'. * 
+ * 
  * @author Misbahul D Munir <misbahuldmunir@gmail.com>
  * @since 1.0
  */
 class Module extends \yii\base\Module
 {
-
+    /**
+     * @inheritdoc
+     */
     public $defaultRoute = 'assigment';
 
     /**
@@ -28,12 +34,17 @@ class Module extends \yii\base\Module
      * @var array
      */
     public $menus;
+    
+    /**
+     * @inheritdoc
+     */
+    public $layout = 'left-menu';
 
     /**
      *
-     * @var string
+     * @var string 
      */
-    public $mainLayout = '@mdm/admin/views/layouts/main.php';
+    private $_mainLayout;
 
     /**
      * Core controller of moduls.
@@ -98,5 +109,59 @@ class Module extends \yii\base\Module
         $this->normalizeController();
 
         return parent::createController($route);
+    }
+
+    /**
+     * 
+     * @return string
+     */
+    public function getMainLayout()
+    {
+        if ($this->_mainLayout === null) {
+            $module = $this->module;
+            while ($module !== null && $module->layout === null) {
+                $module = $module->module;
+            }
+            if ($module !== null && is_string($module->layout)) {
+                $layout = $module->layout;
+            }
+            if (isset($layout)) {
+                if (strncmp($layout, '@', 1) === 0) {
+                    $file = Yii::getAlias($layout);
+                } elseif (strncmp($layout, '/', 1) === 0) {
+                    $file = Yii::$app->getLayoutPath() . DIRECTORY_SEPARATOR . substr($layout, 1);
+                } else {
+                    $file = $module->getLayoutPath() . DIRECTORY_SEPARATOR . $layout;
+                }
+            } else {
+                $file = Yii::getAlias('@app/views/layouts/main');
+            }
+
+            if (pathinfo($file, PATHINFO_EXTENSION) !== '') {
+                return $this->_mainLayout = $file;
+            }
+            $view = Yii::$app->getView();
+            $path = $file . '.' . $view->defaultExtension;
+            if ($view->defaultExtension !== 'php' && !is_file($path)) {
+                $path = $file . '.php';
+            }
+            return $this->_mainLayout = $path;
+        }
+        return $this->_mainLayout;
+    }
+
+    /**
+     * 
+     * @param string $layout
+     */
+    public function setMainLayout($layout)
+    {
+        $file = Yii::getAlias($layout);
+        $view = Yii::$app->getView();
+        $path = $file . '.' . $view->defaultExtension;
+        if ($view->defaultExtension !== 'php' && !is_file($path)) {
+            $path = $file . '.php';
+        }
+        $this->_mainLayout = $path;
     }
 }